pyproject.toml
[tool.poetry]
name = "weather-station"
version = "0.1.0"
description = ""
authors = ["Monique Rio <nique.rio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
flask-socketio = "^5.3.6"
python-dateutil = "^2.8.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
